TableReporter: Remove inheritance in favor of aggregation
Using aggregation here makes the code easier to maintain and moves
towards simplicity and better testability.
